Separate game loop from game UI
